requirements.txt
Django==1.6.1
South==0.8.4
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-autocomplete-light==1.4.14
django-registration==1.0
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.2
rottentomatoes==1.1
static==0.4
wsgiref==0.1.2
